Revert "新增sql字段"
This reverts commit 1f0d6f7a62a361978f49365c3b86b238570e5157.

+++ message-common/src/main/java/com/vv/common/domain/TaskInfo.java
@@ -65,7 +65,7 @@ public class TaskInfo {
     /**
      * 发送账号（邮件下可有多个发送账号、短信可有多个发送账号..）
      */
-    private Long sendAccount;
+    private Integer sendAccount;
 
 
 }
